Disabled the app
I'm sorry to inform you that we have to disable the app for now.
Reason is that the login URL has changed. Therefore, any attempts made to login using the old URL may be reason to flag your account or even ban it. Hopefully, there will be a solution in the next couple of hours.